Move the filters to their own folder
