+++ lib/filters/isSlugMatch.js
