+++ lib/index.js
@@ -5,9 +5,9 @@
  */
 module.exports = function( twigInstance ) {
   return twigInstance.extend( function(){
-    twigInstance.extendFilter('slug', require('./slug'));
-    twigInstance.extendFilter('isSlugMatch', require('./isSlugMatch'));
-    twigInstance.extendFilter('limit', require('./limit'));
-    twigInstance.extendFilter('bustcache', require('./bustcache'));
+    twigInstance.extendFilter('slug', require('./filters/slug'));
+    twigInstance.extendFilter('isSlugMatch', require('./filters/isSlugMatch'));
+    twigInstance.extendFilter('limit', require('./filters/limit'));
+    twigInstance.extendFilter('bustcache', require('./filters/bustcache'));
   });
 };
